sensor: implement unique_id for yaml config
This adds the optional configuration variable unique_id to the sensor
platform. This allows the user to set a unique_id per sensor, the
sensor_type of each entity is added to make it unique per entity.

If this is set, it allows to partially manage the entities from the UI
without enabling the advanced mode. Specifically, this allows to use the
Area setting, which is used in the default Lovelace Dashboard.