sensor: add en and de translation files
Move dewpoint thermal perception strings to translation files and use
keyword constants instead.

Also simplify some states since there meaning gets lost when the
scientific model, especially the boundaries in this case, isn't known
to the user. E.g. remove mentioning of the upper edge which is unknown
to the user.

This fixes #17 since it should be easy for others to add translations
now.