Split SensorThermalComfort
This splits SensorThermalComfort in two separate classes. One
computing/managing class for a set of sensors with the same input
sensors. And a class for the separate sensors.

Computing functions are now decorated so they can be called async in a
synchronized way so that a value is only recalculated once after an update
for all dependent sensors.

This should reduce duplicate calculation and update calls.