Disable config flow entities only once
This changes the config flow entity setup to only set
entity_registry_enabled_default only once. After that it is managed by
the entity registry.

This also changes the selection of the enabled sensor types to a
multichoice. We loose translation for the sensor type itself, but we get
a cleaner and more usable interface for advanced users.

Tests where fixed and some variables renamed to make them more
descriptive.