Drop custom config validator
In preparation of changes in the config processing and error handling
for home-assistant 2023.11 this drops the custom config validator and
exposes the schema through CONFIG_SCHEMA so it is validated from core.

Properly raise exception to notify the user of failure.

see https://developers.home-assistant.io/blog/2023/11/27/config-validation-and-error-handling

fixes #322, fixes #325